Don't update plugins twice
That goes for most of them, at any rate.  `PluginUpdate` gave me a lot
of trouble on the chromebook I didn't see on the other machines, maybe
because it's the old Ubuntu Trusty version.  The command resulted in:

    acer_cb13 | FAILED | rc=1 >>
    GetLatestVimScripts is not vi-compatible; not loaded (you need to set nocp)
    Error detected while processing command line:
    E117: Unknown function: vundle#installer#new

Adding "+set nocp" did nothing.  I tried a lot of things before I noticed
in the man page that '--cmp' args are processed before the vimrc file.
Doesn't work with '-e' either, has to be '-R'